refactor: move files about authentication - 인증, 권한이 관리자, 사용자 모두 사용되기 때문에 공통되게 사용
