+++ survey-api/src/main/java/org/kong/auth/CustomUserDetails.java
@@ -1,9 +1,12 @@
-package org.kong.user.security;
+package org.kong.auth;
 
 import lombok.RequiredArgsConstructor;
+
 import org.kong.user.entity.UserEntity;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
+
+import com.fasterxml.jackson.annotation.JsonIgnore;
 
 import java.util.ArrayList;
 import java.util.Collection;
@@ -11,7 +14,11 @@ import java.util.Collection;
 @RequiredArgsConstructor
 public class CustomUserDetails implements UserDetails {
 
-    private final UserEntity userEntity;
+    private final String userName;
+    @JsonIgnore
+    private final transient String userPwd;
+
+    private String role;
 
     @Override
     public Collection<? extends GrantedAuthority> getAuthorities() {
@@ -19,7 +26,7 @@ public class CustomUserDetails implements UserDetails {
         collection.add(new GrantedAuthority() {
             @Override
             public String getAuthority() {
-                return userEntity.getRole();
+                return role;
             }
         });
         return null;
@@ -27,11 +34,17 @@ public class CustomUserDetails implements UserDetails {
 
     @Override
     public String getPassword() {
-        return userEntity.getUserPwd();
+        return userPwd;
     }
 
     @Override
     public String getUsername() {
-        return userEntity.getUserName();
+        return userName;
+    }
+
+    public CustomUserDetails(UserEntity user) {
+        this.userName = user.getUserName();
+        this.userPwd = user.getUserPwd();
+        this.role = user.getRole();
     }
 }
