+++ survey-api/src/main/java/org/kong/auth/WebConfig.java
@@ -1,4 +1,4 @@
-package org.kong.user.security;
+package org.kong.auth;
 
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
